Refactor: MoveMapDialog
1. devide into UrlParserDialog
2. Refactor MoveMapDialog
