+++ editor/src/MoveMapDialog.tsx
@@ -1,7 +1,7 @@
 import { Dialog, Transition } from "@headlessui/react";
 import { Fragment, useState, useEffect } from "react";
 import { MapController } from "./utils/MapController";
-import { parseMapUrl } from "./utils/MapUrlUtils";
+import UrlParserDialog from "./UrlParserDialog";
 
 type Props = {
     mapController: MapController;
@@ -14,9 +14,6 @@ export default function MoveMapDialog(props: Props): JSX.Element {
     const [coordinates, setCoordinatesState] = useState("");
     const [coordError, setCoordError] = useState("");
     const [isUrlParserOpen, setIsUrlParserOpen] = useState(false);
-    const [urlInput, setUrlInput] = useState("");
-    const [urlError, setUrlError] = useState("");
-    const [parsedResult, setParsedResult] = useState<{ lat: number; lng: number; zoom?: number } | null>(null);
 
     useEffect(() => {
         if (isOpen) {
@@ -68,50 +65,10 @@ export default function MoveMapDialog(props: Props): JSX.Element {
         }
     }
 
-    function handleUrlInputChange(val: string) {
-        setUrlInput(val);
-        if (!val) {
-            setUrlError("");
-            setParsedResult(null);
-            return;
-        }
-
-        const parsed = parseMapUrl(val);
-        if (parsed) {
-            setParsedResult(parsed);
-            setUrlError("");
-        } else {
-            setParsedResult(null);
-            setUrlError("Could not parse URL. Please check the format.");
-        }
-    }
-
-    function handleUrlParserConfirm() {
-        if (parsedResult) {
-            mapController.flyTo(parsedResult.lng, parsedResult.lat, parsedResult.zoom);
-            setIsUrlParserOpen(false);
-            closeModal();
-            setUrlInput("");
-            setUrlError("");
-            setParsedResult(null);
-        } else if (urlInput) {
-            // Try parsing again just in case, or show error if input exists but no result
-            const parsed = parseMapUrl(urlInput);
-            if (parsed) {
-                mapController.flyTo(parsed.lng, parsed.lat, parsed.zoom);
-                setIsUrlParserOpen(false);
-                closeModal();
-                setUrlInput("");
-                setUrlError("");
-                setParsedResult(null);
-            } else {
-                // If invalid, behave like Cancel (close parser, clear state)
-                setIsUrlParserOpen(false);
-                setUrlInput("");
-                setUrlError("");
-                setParsedResult(null);
-            }
-        }
+    function handleUrlParserConfirm(lat: number, lng: number, zoom?: number) {
+        mapController.flyTo(lng, lat, zoom);
+        setIsUrlParserOpen(false);
+        closeModal();
     }
 
     return (
@@ -207,99 +164,11 @@ export default function MoveMapDialog(props: Props): JSX.Element {
             </Transition>
 
             {/* URL Parser Dialog */}
-            <Transition appear show={isUrlParserOpen} as={Fragment}>
-                <Dialog
-                    as="div"
-                    className="fixed inset-0 z-50 overflow-y-auto"
-                    onClose={() => setIsUrlParserOpen(false)}
-                >
-                    <div className="min-h-screen px-4 text-center">
-                        <Transition.Child
-                            as={Fragment}
-                            enter="ease-out duration-300"
-                            enterFrom="opacity-0"
-                            enterTo="opacity-100"
-                            leave="ease-in duration-200"
-                            leaveFrom="opacity-100"
-                            leaveTo="opacity-0"
-                        >
-                            <Dialog.Overlay className="fixed inset-0 bg-black opacity-30" />
-                        </Transition.Child>
-
-                        <span
-                            className="inline-block h-screen align-middle"
-                            aria-hidden="true"
-                        >
-                            &#8203;
-                        </span>
-                        <Transition.Child
-                            as={Fragment}
-                            enter="ease-out duration-300"
-                            enterFrom="opacity-0 scale-95"
-                            enterTo="opacity-100 scale-100"
-                            leave="ease-in duration-200"
-                            leaveFrom="opacity-100 scale-100"
-                            leaveTo="opacity-0 scale-95"
-                        >
-                            <div className="inline-block w-full max-w-md p-6 my-8 overflow-hidden text-left align-middle transition-all transform bg-white shadow-xl rounded-2xl">
-                                <Dialog.Title
-                                    as="h3"
-                                    className="text-lg font-medium leading-6 text-gray-900"
-                                >
-                                    URL Parser
-                                </Dialog.Title>
-                                <div className="mt-4">
-                                    <label className="block text-sm font-medium text-gray-700">
-                                        Paste URL
-                                    </label>
-                                    <input
-                                        type="text"
-                                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm border p-2"
-                                        value={urlInput}
-                                        onChange={(e) => handleUrlInputChange(e.target.value)}
-                                        placeholder="https://..."
-                                    />
-                                    <p className="mt-2 text-xs text-gray-500">
-                                        Supports Google Maps, Apple Maps, OpenStreetMap, Bing Maps.
-                                    </p>
-                                    {parsedResult && (
-                                        <div className="mt-2 text-sm text-green-600">
-                                            Parsed: {parsedResult.lat.toFixed(6)}, {parsedResult.lng.toFixed(6)}
-                                            {parsedResult.zoom !== undefined ? `, Zoom: ${parsedResult.zoom}` : ""}
-                                        </div>
-                                    )}
-                                    {urlError && (
-                                        <p className="mt-2 text-sm text-red-600">
-                                            {urlError}
-                                        </p>
-                                    )}
-                                </div>
-
-                                <div className="mt-6 flex justify-end space-x-3">
-                                    <button
-                                        type="button"
-                                        className="inline-flex justify-center px-4 py-2 text-sm font-medium text-white bg-blue-600 border border-transparent rounded-md hover:bg-blue-700 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
-                                        onClick={handleUrlParserConfirm}
-                                    >
-                                        Confirm
-                                    </button>
-                                    <button
-                                        type="button"
-                                        className="inline-flex justify-center px-4 py-2 text-sm font-medium text-gray-700 bg-white border border-gray-300 rounded-md hover:bg-gray-50 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-indigo-500"
-                                        onClick={() => {
-                                            setIsUrlParserOpen(false);
-                                            setUrlError("");
-                                            setParsedResult(null);
-                                        }}
-                                    >
-                                        Cancel
-                                    </button>
-                                </div>
-                            </div>
-                        </Transition.Child>
-                    </div>
-                </Dialog>
-            </Transition>
+            <UrlParserDialog
+                isOpen={isUrlParserOpen}
+                setIsOpen={setIsUrlParserOpen}
+                onConfirm={handleUrlParserConfirm}
+            />
         </>
     );
 }
